Crypto/aes128-gcm/requirements.txt
pycryptodomex==3.8.1
termcolor
